feat(server): amplify appsync simulator as an extra option
