+++ packages/server/src/index.d.ts
@@ -1,6 +1,10 @@
 import type { YogaServerOptions } from 'graphql-yoga'
+import type { AmplifyAppSyncAPIConfig,  AppSyncMockFile, AppSyncSimulatorDataSourceConfig } from 'amplify-appsync-simulator'
+import type { PrismaAppsyncResolverConfig} from './amplify-appsync-simulator'
+export type { PrismaAppsyncResolverConfig} from './amplify-appsync-simulator'
 
 export const argv: any
+export const argv_appsync_simulator: any
 
 export function createServer({
     schema,
@@ -19,3 +23,14 @@ export function createServer({
 }): void
 
 
+export function createAmplifyAppsyncServer(
+    httpPort: number,
+    lambdaHandler: any,
+    schema: string,
+    resolvers: PrismaAppsyncResolverConfig[],
+    appsyncConfig?: AmplifyAppSyncAPIConfig,
+    additionalMappingTemplates?: AppSyncMockFile[],
+    removeDefaultMappingTemplates?: boolean,
+    additionalDatasources?: AppSyncSimulatorDataSourceConfig[],
+    wsPort?: number
+) :Promise<void>
