+++ packages/server/vtl/lambdaRequestMappingTemplate.vtl
@@ -1,0 +1,5 @@
+{
+    "version": "2018-05-29",
+    "operation": "Invoke",
+    "payload": $util.toJson($context)
+}
