+++ packages/server/vtl/lambdaResponseMappingTemplate.vtl
@@ -1,0 +1,4 @@
+#if($ctx.error)
+     $util.error($ctx.error.message, $ctx.error.type, $ctx.result)
+ #end
+ $util.toJson($ctx.result)
