+++ packages/server/vtl/readVTL.ts
@@ -1,0 +1,6 @@
+import { readFileSync } from 'fs'
+import { resolve } from 'path'
+
+export const readVTL = (filename: string) => {
+    return readFileSync(resolve(__dirname, filename), 'utf8')
+}
